Use an inmutable structure for the state
I'm not particularly proud of this, it could be way simpler.

Blender uses a callback to handle the drawing with a predefined set of
parameters. You can't manually call the function again, which means
the set of parameters have to be mutated to change what the callback
receives. Now couple that with the funky way Python handles the passing of
objects and it's really hard to get something that remains immutable
*and* mutates/is reassinged for the handler. Basically the handler
receives a copy instead of a reference and drawing is never updated.

This is the best solution I have come up with (and the only one that
worked too): put the immutable structures inside a dict, set as a
instance property. Now we can re-assing them (inside the dict) and
have the handler receive the dict (inside a tuple because it's required).
The handler gets a reference to the dict (since it's mutable), so
the new structures show up and drawing gets updated.

¯\_(ツ)_/¯